#1565 Rename monochromator folder to energy
