+++ sgmCommon.pri
@@ -11,22 +11,22 @@ HEADERS += \
     source/beamline/SGM/SGMMAXvMotor.h \
     source/beamline/SGM/SGMBeamline.h \
     source/beamline/SGM/SGMHexapod.h \
-	source/beamline/SGM/SGMHexapodTransformedAxis.h \
-	source/beamline/SGM/monochromator/SGMGratingSupport.h \
-    source/beamline/SGM/monochromator/SGMUndulatorSupport.h \
-	source/beamline/SGM/monochromator/SGMEnergyTrajectory.h \
-    $$PWD/source/beamline/SGM/monochromator/SGMEnergyPosition.h \
-    $$PWD/source/beamline/SGM/monochromator/SGMEnergyControl.h
+    source/beamline/SGM/SGMHexapodTransformedAxis.h \
+    source/beamline/SGM/energy/SGMGratingSupport.h \
+    source/beamline/SGM/energy/SGMUndulatorSupport.h \
+    source/beamline/SGM/energy/SGMEnergyTrajectory.h \
+    source/beamline/SGM/energy/SGMEnergyPosition.h \
+    source/beamline/SGM/energy/SGMEnergyControl.h
 
 
 SOURCES += \
     source/beamline/SGM/SGMMAXvMotor.cpp \
     source/beamline/SGM/SGMBeamline.cpp \
-	source/beamline/SGM/SGMHexapod.cpp \
-	source/beamline/SGM/SGMHexapodTransformedAxis.cpp \
-	source/beamline/SGM/monochromator/SGMEnergyTrajectory.cpp \
-    $$PWD/source/beamline/SGM/monochromator/SGMEnergyPosition.cpp \
-    $$PWD/source/beamline/SGM/monochromator/SGMEnergyControl.cpp
+    source/beamline/SGM/SGMHexapod.cpp \
+    source/beamline/SGM/SGMHexapodTransformedAxis.cpp \
+    source/beamline/SGM/energy/SGMEnergyTrajectory.cpp \
+    source/beamline/SGM/energy/SGMEnergyPosition.cpp \
+    source/beamline/SGM/energy/SGMEnergyControl.cpp
