+++ source/beamline/SGM/energy/SGMEnergyTrajectory.cpp
