+++ source/beamline/SGM/energy/SGMEnergyTrajectory.h
